+++ src/reducers/hasSelfTests.ts
@@ -1,0 +1,28 @@
+import ActionCreator from "../actions";
+
+// This updates a specific search service or collection's self test results property.
+
+const loadCB = (state, action) => {
+  const data = JSON.parse(JSON.stringify(state.data));
+  let dataType = data["search_services"] ? "search_services" : "collections";
+
+  if (!data[dataType].length) {
+    return {};
+  }
+
+  data[dataType].forEach(x => {
+    if (x.id === action.data.self_test_results.id) {
+      x.self_test_results = action.data.self_test_results.self_test_results;
+    }
+  });
+
+  return Object.assign({}, state, {
+    data: data,
+    isFetching: false,
+    isLoaded: true,
+  });
+};
+
+export const extraActions = {
+  [`${ActionCreator.GET_SELF_TESTS}_${ActionCreator.LOAD}`]: loadCB,
+};
