+++ src/reducers/searchServices.ts
@@ -1,31 +1,7 @@
 import { SearchServicesData } from "../interfaces";
+import { extraActions } from "./hasSelfTests";
 import ActionCreator from "../actions";
 import createFetchEditReducer from "./createFetchEditReducer";
-
-// This is only to update a specific search service's self test results property.
-const loadCB = (state, action) => {
-  const data = JSON.parse(JSON.stringify(state.data));
-
-  if (!data.search_services.length) {
-    return {};
-  }
-
-  data.search_services.forEach(search_service => {
-    if (search_service.id === action.data.self_test_results.id) {
-      search_service.self_test_results = action.data.self_test_results.self_test_results;
-    }
-  });
-
-  return Object.assign({}, state, {
-    data: data,
-    isFetching: false,
-    isLoaded: true,
-  });
-};
-
-const extraActions = {
-  [`${ActionCreator.GET_SELF_TESTS}_${ActionCreator.LOAD}`]: loadCB,
-};
 
 export default createFetchEditReducer<SearchServicesData>(
   ActionCreator.SEARCH_SERVICES,
